requirements.txt
Django
psycopg2-binary
django-environ
djangorestframework
Pillow
python-decouple